Camera shtuff
I implemented a simple half circle curve. Need to test ... somehow. Logic is there. If it works, can easily be replicated. Camera is based off of the Orbiting camera code but because I don't understand some of the logic, might just create the camera like how we did it in Camtrans